requirements.txt
openai
groq
python-dotenv
pdfplumber
python-docx
requests
scikit-learn
sentence-transformers
pandas
streamlit